multer img upload

+++ controllers/admin.ts
@@ -2,6 +2,7 @@ import { Request, Response, NextFunction } from "express";
 import Product from "../models/product";
 import { HydratedDocument } from "mongoose";
 import { productType } from "../customTypes";
+import deleteFile from "../middleware/file";
 
 const getAddProduct = (
   req: Request,
@@ -22,10 +23,18 @@ const postAddProduct = (
   next: NextFunction
 ): void => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
   const Id = req.user;
+  if (!image) {
+    res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+    });
+  }
+  const imageUrl = image?.path;
   const product: HydratedDocument<productType> = new Product({
     title: title,
     imageUrl: imageUrl,
@@ -54,7 +63,10 @@ const getEditProduct = (
   const prodId = req.params.productId;
   Product.findById(prodId)
     .then((product) => {
-      if (product === null || product.userId.toString() !== req.user._id.toString()) {
+      if (
+        product === null ||
+        product.userId.toString() !== req.user._id.toString()
+      ) {
         return;
       }
       res.render("admin/edit-product", {
@@ -76,17 +88,24 @@ const postEditProduct = (
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const updatedImage = req.file;
   const updatedDesc = req.body.description;
   Product.findById(prodId)
     .then((product) => {
-      if (product === null || product.userId.toString() !== req.user._id.toString()) {
+      if (
+        product === undefined ||
+        product === null ||
+        product.userId.toString() !== req.user._id.toString()
+      ) {
         return;
       }
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
+      if (updatedImage) {
+        deleteFile(product.imageUrl);
+        product.imageUrl = updatedImage.path;
+      }
       return product.save().then((result) => {
         console.log("UPDATED PRODUCT!");
         res.redirect("/admin/products");
@@ -108,18 +127,25 @@ const getProducts = (req: Request, res: Response, next: NextFunction): void => {
     .catch((err: Error) => console.log(err));
 };
 
-const postDeleteProduct = (
+const postDeleteProduct =  (
   req: Request,
   res: Response,
   next: NextFunction
 ): void => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
+  Product.findById(prodId)
+    .then((product) => {
+      if(product === null){
+        return;
+      }
+      deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+    })
     .then(() => {
       console.log("DESTROYED PRODUCT");
       res.redirect("/admin/products");
     })
-    .catch((err: Error) => console.log(err));
+    .catch((err) => console.log(err));
 };
 
 export {
